move `iter_state_table` to stream common module
