+++ src/stream/src/common/mod.rs
@@ -12,11 +12,26 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod builder;
 pub use builder::*;
-
-mod column_mapping;
 pub use column_mapping::*;
-
-mod infallible_expr;
 pub use infallible_expr::*;
+use risingwave_common::array::Row;
+use risingwave_storage::table::streaming_table::state_table::{RowStream, StateTable};
+use risingwave_storage::StateStore;
+
+use crate::executor::StreamExecutorResult;
+
+mod builder;
+mod column_mapping;
+mod infallible_expr;
+
+pub async fn iter_state_table<'a, S: StateStore>(
+    state_table: &'a StateTable<S>,
+    prefix: Option<&'a Row>,
+) -> StreamExecutorResult<RowStream<'a, S>> {
+    Ok(if let Some(group_key) = prefix {
+        state_table.iter_with_pk_prefix(group_key).await?
+    } else {
+        state_table.iter().await?
+    })
+}
