+++ src/stream/src/executor/managed_state/mod.rs
@@ -16,20 +16,3 @@ pub mod aggregation;
 pub mod dynamic_filter;
 pub mod join;
 pub mod top_n;
-
-use risingwave_common::array::Row;
-use risingwave_storage::table::streaming_table::state_table::{RowStream, StateTable};
-use risingwave_storage::StateStore;
-
-use crate::executor::StreamExecutorResult;
-
-pub async fn iter_state_table<'a, S: StateStore>(
-    state_table: &'a StateTable<S>,
-    prefix: Option<&'a Row>,
-) -> StreamExecutorResult<RowStream<'a, S>> {
-    Ok(if let Some(group_key) = prefix {
-        state_table.iter_with_pk_prefix(group_key).await?
-    } else {
-        state_table.iter().await?
-    })
-}
